chore: Integrate the server code into the repository
The server code has been integrated into the same folder for better
distribution. This integration is doing throught git submodule. Changes
has been did it to the package.json for generating scripts for launch
the server and for deploying the application.

The server could be launch by
    npm run start:dev

And the build process by
    npm run build
or
    gulp